requirements.txt
pandas
pyarrow
tqdm
regex
python-dateutil
pytz
langdetect
matplotlib
numpy
scipy
scikit-learn
snscrape
pyyaml